-n add facilities for annotating JDK classes and annotate Enum
Please read localbin/build-jdk-annotations. You *must* build a personalized
jdk.jar before using the checker, because the jdk.jar is dependent on the
implementation being used. You must rebuild jdk.jar when you change the spec
source, regardless of whether or not you are in Eclipse. There is no automation
for building this because it requires a user parameter.

DO NOT customize version-controlled scripts to automate building for yourself.
Write unversioned scripts so we don't all have to see stuff that only works for
you.